fully use _Float16

+++ src/Native/src/kernels/stackvm/reference/reduce_arg.cpp
@@ -60,7 +60,7 @@ result<void> reduce_arg_impl(TReducer &&reducer, T init_value, const T *input,
             out_map[out_idx].clear();
             out_map[out_idx].push_back(index[axes[0]]);
             dst = src;
-        } else if (std::fabs(src - dst) < epsilon) {
+        } else if (std::fabs(static_cast<float>(src - dst)) < epsilon) {
             out_map[out_idx].push_back(index[axes[0]]);
         }
         return ok();
